Se crea componente tabla
