--- components/customers/index.ts
@@ -1,2 +1,0 @@
-export * from "./DeleteComponent"
-export * from "./QrComponent";
